some smaller changes for flink, and added a new migration test experiment
the test experiment tests whether the migration actually worked, by using a join query where one of the tuples is sent before migration and the other afterward. If the migration fails, i.e., the first part isn't carried over to the new host in a savepoint or snapshot, the second part will not be sufficient to produce output tuples. That means the migration failed. Currently, it succeeds with both Siddhi and Flink